kas/dotnet-nultlib.yml: add lib64 multilib support
As suggested by:
https://github.com/dotnet/core/blob/main/samples/YoctoInstructions.md

While the symlink for /lib64/ld-linux-*.so can be added as a
bbappend, the resulting dotnet binaries were still ending up
with host contamination to the build sysroot. Perhaps another
workaround will solve that issue.

For now, the recommendation is to apply the change either with
kas or in your local.conf, distro.conf or other .conf.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>