dotnet-ld-symlink: add bbclass and recipe
To work around the fact that .NET hard codes the /lib64/ld-linux*.so
path into all binaries, we create a recipe that simply creates the
necessary symlink. The bbclass simply RDEPENDS on the class and also the
packagegroup-dotnet-deps. This simplifies binary-only recipes by only
needing to 'inherit dotnet-ld-symlink'.

Add a symlink to solve the /lib64 issue as suggested by:
https://github.com/dotnet/core/blob/main/samples/YoctoInstructions.md

RDEPENDS on glibc so that /lib/ld-linux-x86-64.so.* is available in
recipe sysroot.

Intentionally restrict to COMPATIBLE_HOST = 'x86_64.*-linux'
Signed-off-by: Tim Orling <tim.orling@konsulko.com>